Create new helper for handling requests
